Fix singleton bug when side stepping make()
Other components can call loadAndCache() directly and this function
should manage the cache itself.  In version 3.1.0 the way singletons
were handled changed, and things in the container managed the cache
outside of loadAndCache().  This fix makes repeated calls to
loadAndCache return the same reference object.  This is the same way
that make works.